data_provider/requirements.txt
boto3>=1.7.31
Pillow>=5.1.0